refactor repository, added usdn proposal
