--- check-onchain-wasm/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "check-onchain-wasm"
-description = "Check the integrity of a WASM code attached to a governance proposal."
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "check-onchain-wasm"
-path = "src/main.rs"
-
-[dependencies]
-clap = { version = "4.4.2", features = ["derive", "env"] }
-tendermint-rpc = { version = "0.40.3", features = ["http-client"]}
-namada_sdk = { git = "https://github.com/anoma/namada", tag = "v0.149.1" }
-namada_governance = { git = "https://github.com/anoma/namada", tag = "v0.149.1" }
-tokio = {version = "1.8.2", default-features = false}
-sha2 = "0.10.8"
