--- increase_target_staked_ratio/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "increase_target_staked_ratio"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-
-[dependencies]
-namada_tx_prelude.workspace = true
-namada_proof_of_stake.workspace = true
-rlsf.workspace = true
-getrandom.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
