--- phase2/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "phase2"
-description = "WASM transaction to transition from phase 1 to phase 2 of Namada mainnet."
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-namada_tx_prelude.workspace = true
-namada_proof_of_stake.workspace = true
-rlsf.workspace = true
-getrandom.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
