--- phase3/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "phase3"
-description = "WASM transaction to transition from phase 2 to phase 3 of Namada mainnet."
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-namada_tx_prelude.workspace = true
-rlsf.workspace = true
-getrandom.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
