--- phase5a/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "phase5a"
-description = "WASM transaction to enable internal transfers of the native token NAM, beginning Phase 5."
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-namada_tx_prelude.workspace = true
-rlsf.workspace = true
-getrandom.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
