+++ proposals/mainnet/add-uusdn-token/data.json
@@ -1,0 +1,13 @@
+{
+  "rust-version": "1.85.1",
+  "proposal": {
+    "title": "Add USDN support to the MASP, whitelist USDN as a gas token, and establish initial rate-limit parameters.",
+    "authors": "Veil",
+    "discussions-to": "https://forum.namada.net/t/proposal-to-add-usdn-support-to-the-masp/1871",
+    "details": "This proposal looks to add support for USDN (Noble’s native yield-bearing stablecoin) to Namada’s Multi-Asset Shielded Pool (MASP), whitelist USDN as a gas token, and set initial rate-limit parameters. Adding USDN would make it the first yield-bearing stable asset in the MASP, enabling shielded deposits that generate real yield and expanding the economic functionality of shielded assets. Whitelisting USDN for gas ensures a seamless user experience, allowing transactions to be paid directly in USDN. The initial parameters are a ___M global mint limit, ___M per-epoch limit, and a gas price of ___ uusdn per gas unit.",
+    "author": "tnam1qq92nknz9u82aewcw3k0nrp7t7d5ytgskcjkyduk",
+    "voting_start_epoch": 6,
+    "voting_end_epoch": 12,
+    "activation_epoch": 18
+  }
+}
