+++ proposals/mainnet/update-staking-inflation-parameters/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "update-staking-inflation-parameters"
+description = "WASM transaction to update staking inflation parameters."
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+
+[dependencies]
+namada_tx_prelude_01491.workspace = true
+rlsf.workspace = true
+getrandom.workspace = true
+
+[lib]
+crate-type = ["cdylib"]
