+++ proposals/template/update-wasm/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "update-wasm"
+description = "WASM transaction template to update a wasm transaction."
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+namada_tx_prelude_01491.workspace = true
+rlsf.workspace = true
+getrandom.workspace = true
+
+[lib]
+crate-type = ["cdylib"]
