+++ proposals/template/update-wasm/src/tx.b74104949ac0c35ee922fdc3f3db454627742e2483d79550c12fcf31755c6d01.wasm
@@ -1,0 +1,1 @@
+placeholder
