+++ proposals/testnet/add-uusdc-gas-token/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "add-uusdc-gas-token"
+description = "WASM transaction to allow transfer/channel-27/uusdc as a gas token on housefire testnet."
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+namada_tx_prelude_01491.workspace = true
+rlsf.workspace = true
+getrandom.workspace = true
+
+[lib]
+crate-type = ["cdylib"]
