--- update_ibc_rate_limits/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "update_ibc_rate_limits"
-description = "WASM transaction to change IBC rate limits."
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-namada_ibc.workspace = true
-namada_tx_prelude.workspace = true
-rlsf.workspace = true
-getrandom.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
